+++ __mocks__/cleanStackMock.js
@@ -1,0 +1,5 @@
+function cleanStack(stack) {
+  return stack;
+}
+module.exports = cleanStack;
+module.exports.default = cleanStack;