+++ packages/decap-cms-backend-forgejo/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.2](https://github.com/decaporg/decap-cms/compare/decap-cms-backend-forgejo@3.4.1...decap-cms-backend-forgejo@3.4.2) (2026-09-10)
+
+### Bug Fixes
+
+- support Forgejo editorial workflow PR heads ([#7976](https://github.com/decaporg/decap-cms/issues/7976)) ([5984304](https://github.com/decaporg/decap-cms/commit/59843046caf6cd9b6b95b6744c5e7536f8045611))
+
+## [3.4.1](https://github.com/decaporg/decap-cms/compare/decap-cms-backend-forgejo@3.4.0...decap-cms-backend-forgejo@3.4.1) (2026-09-08)
+
+**Note:** Version bump only for package decap-cms-backend-forgejo
+
 # 3.4.0 (2026-07-23)
 
 **Note:** Version bump only for package decap-cms-backend-forgejo